De-duplicate code between Video and Audio content models.
